Create solution for Exercise 12 findTheOldest
Spent way too much time trying to get a more concise and modern solution,
in the end I stuck with what I know can work. Looking at the solution
branch on github, I can see now how it works, but it is still not
intuitive how I'm supposed to get to that solution.